# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaravelVueErp API",
        "description": "API Documentation for LaravelVueErp Project Management System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/projects/enquiries/{enquiry}": {
            "get": {
                "tags": [
                    "Enquiries"
                ],
                "summary": "Get enquiry details",
                "operationId": "d76fbf2d3734fa52a2ce9baf01320c42",
                "parameters": [
                    {
                        "name": "enquiry",
                        "in": "path",
                        "description": "Enquiry ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Enquiry details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProjectEnquiry"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Enquiry not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Enquiries"
                ],
                "summary": "Update enquiry details",
                "operationId": "4f70f43f68171b7eededeb947f0034b5",
                "parameters": [
                    {
                        "name": "enquiry",
                        "in": "path",
                        "description": "Enquiry ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Office Branding Project"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated project description"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "pending",
                                            "in_progress",
                                            "quote_approved",
                                            "converted_to_project",
                                            "cancelled"
                                        ]
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high",
                                            "urgent"
                                        ]
                                    },
                                    "estimated_budget": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "expected_delivery_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "contact_person": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "assigned_po": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Enquiry updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ProjectEnquiry"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Enquiry not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Enquiries"
                ],
                "summary": "Delete enquiry",
                "operationId": "8fa95306f3e8c100f1cf8cbeb987c381",
                "parameters": [
                    {
                        "name": "enquiry",
                        "in": "path",
                        "description": "Enquiry ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Enquiry deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Enquiry not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projects/enquiries/{enquiry}/approve-quote": {
            "post": {
                "tags": [
                    "Enquiries"
                ],
                "summary": "Approve enquiry quote",
                "operationId": "e106006e3bb04b5eb234d84b1e8f314b",
                "parameters": [
                    {
                        "name": "enquiry",
                        "in": "path",
                        "description": "Enquiry ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quote approved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ProjectEnquiry"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Enquiry not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProjectEnquiry": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "enquiry_number": {
                        "type": "string",
                        "example": "ENQ-2024-0001"
                    },
                    "title": {
                        "type": "string",
                        "example": "Office Branding Project"
                    },
                    "description": {
                        "type": "string",
                        "example": "Complete branding solution for new office"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "pending",
                            "in_progress",
                            "quote_approved",
                            "converted_to_project",
                            "cancelled"
                        ]
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high",
                            "urgent"
                        ]
                    },
                    "client_id": {
                        "type": "integer"
                    },
                    "contact_person": {
                        "type": "string",
                        "example": "John Smith"
                    },
                    "estimated_budget": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "date_received": {
                        "type": "string",
                        "format": "date"
                    },
                    "expected_delivery_date": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}